pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-chart-analysis"
version = "0.1.0"
description = "자동 차트 분석 프로젝트"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "joju" }]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "ta>=0.10.0",
    "yfinance>=0.2.28",
    "finance-datareader>=0.9.50",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
    "google-api-python-client>=2.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0", "black>=23.0.0", "ruff>=0.1.0"]
notebook = ["jupyter>=1.0.0", "ipykernel>=6.25.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
